+++ webpack.config.js
@@ -1,0 +1,50 @@
+const path = require('path');
+
+module.exports = {
+    entry: './src/js/app.js',
+    output: {
+        path: path.resolve(__dirname, 'sickrage/core/webserver/static/js'),
+        filename: 'core.min.js'
+    },
+    module: {
+        rules: [
+            {
+                test: /\.js$/,
+                exclude: /node_modules/,
+                use: {
+                    loader: "babel-loader"
+                }
+            },
+            {
+                test: /\.(png|woff|woff2|eot|ttf|svg)$/,
+                use: [{
+                    loader: 'file-loader',
+                    options: {
+                        name: '[name].[ext]',
+                        outputPath: '../fonts/'
+                    }
+                }]
+            },
+            {
+                test: /\.(scss)$/,
+                use: [{
+                    loader: 'style-loader' // inject CSS to page
+                }, {
+                    loader: 'css-loader' // translates CSS into CommonJS modules
+                }, {
+                    loader: 'postcss-loader', // Run post css actions
+                    options: {
+                        plugins: function () { // post css plugins, can be exported to postcss.config.js
+                            return [
+                                require('precss'),
+                                require('autoprefixer')
+                            ];
+                        }
+                    }
+                }, {
+                    loader: 'sass-loader' // compiles Sass to CSS
+                }]
+            }
+        ]
+    }
+};