+++ changelog.md
@@ -1,6 +1,7 @@
 # Changelog
 
-- * 8e6b694 - 2018-07-09: Compiled all CSS/SCSS into 1 core file 
+- * 3374cb5 - 2018-07-09: Migrating away from Bower to NPM/WebPack 
+- * 86021ad - 2018-07-09: Compiled all CSS/SCSS into 1 core file 
 - * b912b2e - 2018-07-08: Refactored config views to Bootstrap 4.x 
 - * ae015ef - 2018-07-08: Release v9.3.54 
 - * afa499a - 2018-07-08: Set secure cookie to expire after 30 days